==> notebooks/2-kialo/1_kialo_scraping.ipynb <==
> **Note:** This notebook's code is heavily inspired by [this Kialo tool implementation](https://github.com/4mbroise/ADC/blob/main/tool/kialo.ipynb).

from dotenv import load_dotenv
import os

load_dotenv()
kialoUsername = os.getenv("KIALO_USERNAME")
secret = os.getenv("KIALO_PASSWORD")

if not kialoUsername or not secret:
    raise ValueError("Username/password missing! Check the .env file at the root of the project.")

from enum import Enum
import requests
import pandas as pd
import os

class KialoSort(Enum):
  RANK_ACTIVITY = "rank_and_latest_activity"
  VIEW          = "view_count"
  LAST_ACTIVITY = "latest_activity"

class KialoFilter(Enum):
  PROMOTED      = "promoted"
  PARTICIPATE   = "participate"
  LAST_ACTIVITY = "latest_activity"
  TAG           = "tag"
  TAG_ALL       = "tag_all"

class KialoTool:
  def getTags(self):
    tags = requests.get("https://www.kialo.com/api/v1/discussiontags")
    return [item['tagName'] for item in tags.json()["tags"]]

  def getDiscussions(self, filter: KialoFilter, sort: KialoSort, limit=500): # Scrape 500 discussions
    req = "https://www.kialo.com/api/v1/discussions?filter=" + str(filter.value) + "&sort=" + str(sort.value) + "&limit=" + str(limit) + "&skip=0"
    tags = requests.get(req)
    return tags.json()["discussions"]

  def discussions2urlID(self, discussions, export=False):
    idsUrl = [x["title"].lower().replace("?", "").replace(" ", "-")+"-"+str(x["id"]) for x in discussions]
    tags = [x["tags"] for x in discussions]
    if export:
      pd.DataFrame.from_dict({"kialoUrlId" : idsUrl, "tags" : tags}).to_csv("../../data/kialo/kialo-url-ids.csv")

    return idsUrl

k = KialoTool()

k.discussions2urlID(k.getDiscussions(filter=KialoFilter.TAG, sort=KialoSort.RANK_ACTIVITY), export=True)

from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By
import time
from tqdm import tqdm

def downloadDiscussions(disscusionUrlIds):
  path = os.path.abspath("../../data/kialo/debates")

  prefs = {"download.default_directory": path }
  options = Options()
  prefs = {"download.default_directory": os.path.abspath("../../data/kialo/debates")}
  options = Options()
  options.add_argument("--headless")  # no graphical interface
  options.add_argument("--no-sandbox")
  options.add_argument("--disable-dev-shm-usage")
  options.add_argument("--disable-gpu")
  options.add_argument("--window-size=1920,1080")
  options.add_experimental_option("prefs", prefs)

  driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()), options=options)

  driver.get("https://www.kialo.com/login")

  #Login
  id          = driver.find_element(By.ID, "emailOrUsername")
  password    = driver.find_element(By.ID, "password")
  loginButton = driver.find_element(By.CLASS_NAME, "login-form__submit")
  id.send_keys(kialoUsername)
  password.send_keys(secret)

  loginButton.click()

  WebDriverWait(driver, 10).until(
    EC.presence_of_element_located(
        (By.CLASS_NAME, "portal-nav-item-for-page__button--explore")
    )
)

  for urlId in tqdm(disscusionUrlIds):
    driver.get("https://www.kialo.com/export/" + urlId + ".txt")
    time.sleep(1.5)

downloadDiscussions(k.discussions2urlID(k.getDiscussions(filter=KialoFilter.TAG, sort=KialoSort.RANK_ACTIVITY), export=True))
